+++ src/Bridges/NetteDI/MyTesterExtension.php
@@ -7,7 +7,7 @@ use Exception;
 use MyTester\Bridges\NetteRobotLoader\TestSuitesFinder;
 use MyTester\CodeCoverage\Collector;
 use MyTester\CodeCoverage\Helper as CodeCoverageHelper;
-use MyTester\CodeCoverage\PercentFormatter;
+use MyTester\CodeCoverage\Formatters\PercentFormatter;
 use MyTester\Tester;
 use Nette\DI\Helpers;
 use Nette\Schema\Expect;